+++ src/main/kotlin/com/example/kbocombo/crawler/client/KboClient.kt
@@ -1,10 +1,8 @@
-package com.example.kbocombo.crawler
+package com.example.kbocombo.crawler.client
 
-import org.springframework.http.HttpHeaders
 import org.springframework.http.ResponseEntity
 import org.springframework.util.MultiValueMap
 import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.service.annotation.HttpExchange
 import org.springframework.web.service.annotation.PostExchange
