+++ src/main/kotlin/com/example/kbocombo/crawler/client/KboClientConfig.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.crawler
+package com.example.kbocombo.crawler.client
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
