+++ src/main/kotlin/com/example/kbocombo/crawler/service/RegisteredPlayerReader.kt
@@ -1,5 +1,6 @@
-package com.example.kbocombo.crawler
+package com.example.kbocombo.crawler.service
 
+import com.example.kbocombo.crawler.client.KboClient
 import com.example.kbocombo.entity.PlayerPosition
 import com.example.kbocombo.entity.Team
 import com.example.kbocombo.entity.WebId
